Fire-Emblem-67-LT-Editor/requirements_dev.txt
pygame-ce==2.3.2
pyinstaller==6.2.0
typing-extensions==4.8.0
PyQt5==5.15.10
mypy==1.8.0
mypy-extensions==1.0.0

# for compiling docs
attrs>=21.4.0
sphinx_rtd_theme==2.0.0
myst-parser==2.0.0